chore: move demo helper files out of the expo-router routes tree
expo-router treats every file under app/ as a route and warned at startup that the four shared helpers under app/demo/lib/ (AnimatedTrendTextInput, DemoScreen, shared, styles) are "missing the required default export" — they're imported by the demo screens, not navigable routes. Those 4 warnings fired the "Open debugger to view warnings" LogBox toast on every screen.

Move app/demo/lib/ -> demo-lib/ (a sibling of app/, outside the routing tree) and update the 49 relative imports (./lib/X -> ../../demo-lib/X); shared.ts's sim imports go ../../../sim/ -> ../sim/. expo-router only special-cases +-prefixed files, _layout, and (groups), and v6 exposes no app.json ignore for arbitrary non-route files, so relocating is the canonical fix.

Verified on the iOS 26 sim: home + demo screens load with no warning toast; Playground (imports all four helpers incl. the AnimatedTrendTextInput header readout) renders correctly. typecheck + lint + 652 tests pass.
