+++ app/demo/edge-cases.tsx
@@ -5,9 +5,9 @@ import { LiveChart } from "react-native-livechart";
 import { useSharedValue } from "react-native-reanimated";
 
 import { useSimulatedChartData } from "../../sim/useSimulatedChartData";
-import { DemoScreen } from "./lib/DemoScreen";
-import { ACCENT } from "./lib/shared";
-import { demoStyles } from "./lib/styles";
+import { DemoScreen } from "../../demo-lib/DemoScreen";
+import { ACCENT } from "../../demo-lib/shared";
+import { demoStyles } from "../../demo-lib/styles";
 
 export const options = { title: "Empty, loading, formatters" };
