+++ app/demo/historical-data.tsx
@@ -3,9 +3,9 @@ import { Pressable, Text, View } from "react-native";
 import { LiveChart, type LiveChartPoint } from "react-native-livechart";
 import { useSharedValue } from "react-native-reanimated";
 
-import { DemoScreen } from "./lib/DemoScreen";
-import { ACCENT } from "./lib/shared";
-import { demoStyles } from "./lib/styles";
+import { DemoScreen } from "../../demo-lib/DemoScreen";
+import { ACCENT } from "../../demo-lib/shared";
+import { demoStyles } from "../../demo-lib/styles";
 
 export const options = { title: "Historical data fill" };
