+++ app/demo/momentum.tsx
@@ -4,9 +4,9 @@ import type { Momentum, MomentumConfig } from "react-native-livechart";
 import { LiveChart } from "react-native-livechart";
 
 import { useSimulatedChartData } from "../../sim/useSimulatedChartData";
-import { DemoScreen } from "./lib/DemoScreen";
-import { ACCENT, VOLATILITY_MODES } from "./lib/shared";
-import { demoStyles } from "./lib/styles";
+import { DemoScreen } from "../../demo-lib/DemoScreen";
+import { ACCENT, VOLATILITY_MODES } from "../../demo-lib/shared";
+import { demoStyles } from "../../demo-lib/styles";
 
 export const options = { title: "Momentum" };
