+++ demo-lib/AnimatedTrendTextInput.tsx
